waypoint_selection_ui/requirements.txt
fastapi==0.131.0
uvicorn==0.34.3
jinja2==3.1.6
pydantic==2.10.6
python-multipart==0.0.22
numpy==2.2.4
pyyaml==6.0.3
httpx==0.28.1